tests/wpt/tests/layout-instability/local-shift-without-viewport-shift-2.html

<!DOCTYPE html>
<title>Layout Instability: local shift without viewport shift</title>
<link rel="help" href="https://wicg.github.io/layout-instability/" />
<style>

#c { position: relative; width: 300px; height: 100px; scale: 0.1; }
#j { position: relative; width: 100px; height: 10px; background: blue; }

</style>
<div id='c'>
  <div id='j'></div>
</div>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/util.js"></script>
<script>

promise_test(async () => {
  const watcher = new ScoreWatcher;

  // Wait for the initial render to complete.
  await waitForAnimationFrames(2);

  document.querySelector("#j").style.top = "4px";

  // Make sure no shift score is reported, since the element didn't move in the
  // viewport.
  await waitForAnimationFrames(3);
  assert_equals(watcher.score, 0);
}, "Local shift without viewport shift.");

</script>
